Adjust timeouts when fetching robots.txt
The options OVERRIDE_CONNECT_TIMEOUT and OVERRIDE_READ_TIMEOUT
where only applied during retrieve_head, and not retrieve_page.
As a result, the code to fetch /robots.txt was using the
default timeout of 3 minutes.  Fixed that.

Also, the retry logic was explicitly .start-ing the connections
and not re-setting the timeouts.  As a result, the timeouts only
applied to the first attempt.

Updated the description of these two parameters in config-advanced.